Task #3168: Merged trunk with task branch

+++ branches/MAC-8bitmode-Task3168/CEP/GSM/bremen/src/pipeline.py
@@ -1,10 +1,9 @@
 #!/usr/bin/env python
 import monetdb.sql as db
-from src.errors import SourceException
+from src.errors import SourceException, ImageStateError
 from src.gsmconnectionmanager import GSMConnectionManager
 from src.gsmlogger import get_gsm_logger
-from src.queries import get_insert_temprunningcatalog
-from src.sqllist import get_sql
+from src.sqllist import get_sql, get_svn_version
 from src.grouper import Grouper
 from src.updater import run_update
 import logging
@@ -40,6 +39,18 @@ class GSMPipeline(object):
             raise exc
         self.log.info('Pipeline started.')
 
+    def reopen_connection(self):
+        """
+        Reopen connection in case it was closed.
+        """
+        if not self.conn or not self.conn.established():
+            try:
+                self.conn = self.conn_manager.get_connection(**params)
+                self.log.info('Pipeline connection reopened.')
+            except db.Error as exc:
+                self.log.error("Failed to connect: %s" % exc)
+                raise exc
+
     def read_image(self, source):
         """
         Read image and detections from a given source.
@@ -72,34 +83,56 @@ class GSMPipeline(object):
             grouper.cleanup()
         self.conn.execute(get_sql("GroupFill"))
 
-    def process_image(self, image_id):
+    def process_image(self, image_id, sources_loaded=False):
         """
         Process single image.
+        @sources_loaded: True if there are records in the extractedsources
+        already.
         """
         self.conn.start()
+        status, band, stokes = self.conn.exec_return("""
+        select status, band, stokes
+          from images
+         where imageid = %s;""" % image_id, single_column=False)
+        if status == 1:
+            raise ImageStateError('Image %s in state 1 (Ok). Cannot process' %
+                                  image_id)
         self.conn.execute("delete from temp_associations;")
-        self.conn.execute(get_sql('insert_extractedsources', image_id))
-        self.conn.execute(get_insert_temprunningcatalog(image_id, 1.0))
-        self.conn.execute_set(get_sql('Associate extended',
+        if not sources_loaded:
+            self.conn.execute(get_sql('insert_extractedsources', image_id))
+            self.conn.execute(get_sql('insert dummysources', image_id))
+        self.conn.execute(get_sql('Associate point',
                                       image_id, math.sin(0.025), 1.0))
+        self.conn.execute_set(get_sql('Associate extended',
+                                      image_id, math.sin(0.025), 0.5))
         self.conn.call_procedure("fill_temp_assoc_kind();")
         # Process one-to-one associations;
         self.conn.execute(get_sql('add 1 to 1'))
         #process one-to-many associations;
         self.conn.execute(get_sql('add 1 to N'))
-        #process one-to-many associations;
-        self.conn.execute_set(get_sql('add N to 1'))
+        self.conn.execute_set(get_sql('update flux_fraction'))
+        #process many-to-one associations;
+        self.conn.execute_set(get_sql('add N to 1', band))
         #Process many-to-many;
         self.run_grouper()
 
-        band = self.conn.exec_return("select band from images"\
-                                     " where imageid = %s;" % image_id)
         #updating runningcatalog
-        run_update(self.conn, 'PG update runningcatalog', image_id)
+        run_update(self.conn, 'update runningcatalog', image_id)
+        run_update(self.conn, 'update runningcatalog extended', image_id)
+        self.conn.execute(get_sql('update runningcatalog XYZ', image_id))
         #First update, then insert new (!!!)
-        run_update(self.conn, 'PG update runningcatalog_fluxes',
-                                  image_id, band)
-        self.conn.execute(get_sql('PG insert new bands', image_id, band))
+        run_update(self.conn, 'update runningcatalog_fluxes',
+                             image_id)
+        self.conn.execute(get_sql('insert new bands for point sources',
+                                  image_id, band))
         #inserting new sources
         self.conn.execute_set(get_sql('Insert new sources', image_id))
+        self.conn.execute_set(get_sql('Join extended', image_id))
+        #update image status and save current svn verion.
+        self.conn.execute("""
+update images
+   set status = 1,
+       process_date = current_timestamp,
+       svn_version = %s
+ where imageid = %s""" % (get_svn_version(), image_id))
         self.conn.commit()
